Refactor: move append_blank_log() to ReplicationHandler
